refactor(checker): route constraint indexed access relation guard
