+++ crates/tsz-checker/src/checkers/generic_checker/mod.rs
@@ -1294,6 +1294,7 @@ mod array_like_constraint_helpers;
 mod callable_constraint_helpers;
 mod conditional_constraint_helpers;
 mod constraint_display_helpers;
+mod constraint_indexed_access_helpers;
 mod constraint_syntax_instantiation;
 mod constraint_validation;
 mod constructor_accessibility_helpers;
