+++ scripts/arch/arch_guard_config.py
@@ -755,6 +755,13 @@ REGEX_LINE_COUNT_CHECKS = [
             / "crates"
             / "tsz-checker"
             / "src"
+            / "checkers"
+            / "generic_checker"
+            / "constraint_indexed_access_helpers.rs",
+            ROOT
+            / "crates"
+            / "tsz-checker"
+            / "src"
             / "types"
             / "computation"
             / "complex_contextual_new.rs",
